Move icons to shared code
