+++ shared/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@dumbcode/shared",
+  "main": "index.ts",
+  "version": "0.0.0",
+  "devDependencies": {
+    "@types/react": "^18.0.18"
+  },
+  "dependencies": {
+    "react": "^18.2.0"
+  }
+}
