+++ studio/src/components/CommandInputBar.tsx
@@ -1,9 +1,9 @@
+import { SVGTerminal } from "@dumbcode/shared/icons";
 import { MouseEvent, useEffect, useLayoutEffect, useRef, useState } from "react";
 import { useStudio } from "../contexts/StudioContext";
 import { Command } from "../studio/command/Command";
 import { CommandRoot, ParsedArgument } from "../studio/command/CommandRoot";
 import { useListenableObjectNullable } from "../studio/listenableobject/ListenableObject";
-import { SVGTerminal } from "./Icons";
 
 const CommandInputBar = ({ command }: { command?: CommandRoot }) => {
   const { getSelectedProject, onFrameListeners } = useStudio()
