+++ studio/src/dialogboxes/ExportAnimationAsGifDialogBox.tsx
@@ -1,9 +1,9 @@
+import { SVGDownload, SVGInfoBubble } from "@dumbcode/shared/icons"
 import Image from "next/image"
 import { useCallback, useEffect, useRef, useState } from "react"
 import InputSlider from "react-input-slider"
 import { Color } from "three"
 import Checkbox from "../components/Checkbox"
-import { SVGDownload, SVGInfoBubble } from "../components/Icons"
 import { DisplayModeDropup, RenderModeDropup, TextureGroupDropup } from "../components/InfoBar"
 import NumericInput from "../components/NumericInput"
 import { RawCanvas } from "../components/StudioCanvas"
